% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReConPlot.R
\docType{data}
\name{cn_data}
\alias{cn_data}
\title{Test copy number data}
\format{
A data frame with 2223 rows and 5 variables
}
\usage{
data(cn_data)
}
\description{
A dataset containing an example data frame with
2223 CN segments. The variables are as follows:
}
\details{
\itemize{
\item chr. Chromosome
\item start. Start
\item end. End
\item copyNumber. Total copy number
\item minorAlleleCopyNumber.  Minor allele copy number
}
}
\keyword{datasets}
