% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement_plot.R
\name{rearrangement_plot}
\alias{rearrangement_plot}
\title{Function to plot genomic rearrangements (allele-specific copy number profiles and structural variants)}
\usage{
rearrangement_plot(
  sv,
  cnv,
  title = "",
  genes = NULL,
  chr_selection = NULL,
  scaling_cn_SVs = 1/6,
  scale_separation_SV_type_labels = 1/12,
  pos_SVtype_description = 10000000,
  window = 10000000,
  xscale = 10 * 10^6,
  percentage_increase_y_axis = 0.1,
  legend_SV_types = TRUE,
  size_chr_labels = 7,
  size_title = 7,
  colour_band2 = "antiquewhite1",
  colour_band1 = "grey86",
  colour_DEL = "orange",
  colour_h2hINV = "forestgreen",
  colour_DUP = "darkblue",
  colour_t2tINV = "black",
  size_gene_label = 2.2,
  color_minor_cn = "darkred",
  curvature_intrachr_SVs = -0.3,
  curvature_interchr_SVs = -0.35,
  max.cnv = 10,
  npc_now = 0.00625,
  scale_ticks = 20000000
)
}
\arguments{
\item{sv}{Dataframe with SV information. Required, no default value.}

\item{cnv}{Dataframe with copy number information. Required, no default value.}

\item{title}{Title of the plot. Defaults to "".}

\item{genes}{Vector of genes (HUGO gene names) to be shown on the plot. Defaults to NULL.}

\item{chr_selection}{Chromosomes, start and end positions to be depicted. Defaults to NULL.}

\item{scaling_cn_SVs}{Relative dimension of the panel representing the SVs with respect to the copy number profile XX. Defaults to 1/6.}

\item{scale_separation_SV_type_labels}{Separatation (denoted as a fraction of the y axis XXX) between the SV labels/legend. Defaults to 1/12.}

\item{pos_SVtype_description}{Position for the SV labels/legend on the y axis. Defaults to 10000000.}

\item{window}{extra spacing on the x axis around the leftmost and rightmost breakpoints detected in the chromosomes selected unless a specific start and end positions for the plot are input.}

\item{xscale}{Scale for the x axis. Defaults to 10*10^6 to put the x axis in Mbp.}

\item{percentage_increase_y_axis}{Relative percentage to increase the distance of the y axis. XX}

\item{legend_SV_types}{Whether to show the legend for the SV types or not. Defaults to TRUE.}

\item{size_chr_labels}{Size of the chromsosome labels. Defaults to 7pt (use 5-7pt for publication-ready figures).}

\item{size_title}{Size of the plot title. Defaults to 7pt (use 5-7pt for publication-ready figures).}

\item{colour_band2}{Colour of the second background horizintal stripe. Defaults to "antiquewhite1"}

\item{colour_band1}{Colour of the first background horizintal stripe. Defaults to "grey86"}

\item{colour_DEL}{Colour of the arcs representing deletions (DEL)}

\item{colour_h2hINV}{Colour of the arcs representing head-to-head inversions (h2hINV).}

\item{colour_DUP}{Colour of the arcs representing duplications (DUP).}

\item{colour_t2tINV}{Colour of the arcs representing tail-to-tail inversions (t2tINV).}

\item{size_gene_label}{Size of the gene labels. Defaults to 2.2.}

\item{color_minor_cn}{Colour for the horizontal bars representing the minor copy number values.}

\item{curvature_intrachr_SVs}{Curvature for the arcs represeting intrachromosomal SVs. Defaults to -0.3}

\item{curvature_interchr_SVs}{Curvature for the arcs represeting interchromosomal SVs. Defaults to -0.35}

\item{max.cnv}{Cap on the total copy number (for the minor we do not need as the minor will never be very high). Defaults to 10.}

\item{scale_ticks}{Spacing of breaks in the x axis (in bp). Defaults to 20000000 (i.e., 20Mb).}
}
\description{
This function generates publication-quality plots for copy number and structural variation profiles, which are particularly useful in the context of cancer genome analysis projects.
}
\examples{
TODO
}
