% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReConPlot.R
\docType{data}
\name{snv_data}
\alias{snv_data}
\title{Test single nucleotide variation data}
\format{
A data frame with 100 rows and 3 variables
}
\usage{
data(snv_data)
}
\description{
A dataset containing an example data frame with
100 SNVs to test the custom_annotation module. The variables are as follows:
}
\details{
\itemize{
\item chr. Chromosome
\item pos. Position
\item y. Y coordinate (in this case, variant allele frequency)
}
}
\keyword{datasets}
