% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReConPlot.R
\docType{data}
\name{sv_data}
\alias{sv_data}
\title{Test structural variation data}
\format{
A data frame with 538 rows and 5 variables
}
\usage{
data(sv_data)
}
\description{
A dataset containing an example data frame with
538 SVs. The variables are as follows:
}
\details{
\itemize{
\item chr1. Chromosome 1 of the SV
\item pos1. Position 1 of the SV
\item chr2. Chromosome 2 of the SV
\item pos2. Position 2 of the SV
\item strands. Orientation of the SV (++, --, +- or -+)
}
}
\keyword{datasets}
