% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReConPlot.R
\name{ReConPlot}
\alias{ReConPlot}
\title{Function to plot genomic rearrangements (allele-specific copy number profiles and structural variants)}
\usage{
ReConPlot(
  sv,
  cnv,
  title = "",
  genes = NULL,
  chr_selection = NULL,
  scaling_cn_SVs = 1/5,
  scale_separation_SV_type_labels = 1/22,
  pos_SVtype_description = 5000000,
  window = 10000000,
  xscale = 10 * 10^6,
  percentage_increase_y_axis = 0.1,
  legend_SV_types = TRUE,
  size_chr_labels = 7,
  size_title = 7,
  size_text = 5,
  size_gene_label = 1.5,
  colour_band1 = "grey86",
  colour_band2 = "antiquewhite1",
  colour_DEL = "orange",
  colour_h2hINV = "forestgreen",
  colour_DUP = "darkblue",
  colour_t2tINV = "black",
  colour_TRA = "darkgray",
  color_minor_cn = "#8491B4B2",
  upper_limit_karyotype = -0.2,
  karyotype_rel_size = 0.2,
  custom_annotation = NULL,
  ann_dot_col = "black",
  ann_dot_size = 0.5,
  ann_y_title = "",
  ann_rel_size = 0.4,
  curvature_intrachr_SVs = -0.15,
  curvature_interchr_SVs = -0.08,
  max.cn = 8,
  npc_now = 0.00625 * 3,
  scale_ticks = 20000000,
  size_interchr_SV_tip = 0.2,
  label_interchr_SV = FALSE,
  size_sv_line = 0.1,
  genome_version = "hg38"
)
}
\arguments{
\item{sv}{Dataframe with SV information. Required, no default value.}

\item{cnv}{Dataframe with copy number information. Required, no default value.}

\item{title}{Title of the plot. Defaults to "".}

\item{genes}{Vector of genes (HUGO gene names) to be shown on the plot. Defaults to NULL.}

\item{chr_selection}{Chromosomes, start and end positions to be depicted. Defaults to NULL.}

\item{scaling_cn_SVs}{Relative dimension of the panel representing the SVs with respect to the copy number profile. Defaults to 1/6.}

\item{scale_separation_SV_type_labels}{Separatation (denoted as a fraction of the y axis XXX) between the SV labels/legend. Defaults to 1/18.}

\item{pos_SVtype_description}{Position for the SV labels/legend on the y axis. Defaults to 1000000.}

\item{window}{extra spacing on the x axis around the leftmost and rightmost breakpoints detected in the chromosomes selected unless a specific start and end positions for the plot are input.}

\item{xscale}{Scale for the x axis. Defaults to 10*10^6 to put the x axis in Mbp.}

\item{percentage_increase_y_axis}{Relative percentage to increase the distance of the y axis. XX}

\item{legend_SV_types}{Whether to show the legend for the SV types or not. Defaults to TRUE.}

\item{size_chr_labels}{Size of the chromsosome labels. Defaults to 7pt (use 5-7pt for publication-ready figures).}

\item{size_title}{Size of the plot title. Defaults to 7pt (use 5-7pt for publication-ready figures).}

\item{size_text}{Size of the plot text. Defaults to 5pt (use 5-7pt for publication-ready figures).}

\item{size_gene_label}{Size of the gene label text. Defaults to 1.5.}

\item{colour_band1}{Colour of the first background horizontal stripe. Defaults to "grey86".}

\item{colour_band2}{Colour of the second background horizontal stripe. Defaults to "antiquewhite1".}

\item{colour_DEL}{Colour of the arcs representing deletions (DEL). Defaults to "orange".}

\item{colour_h2hINV}{Colour of the arcs representing head-to-head inversions (h2hINV). Defaults to "forestgreen".}

\item{colour_DUP}{Colour of the arcs representing duplications (DUP). Defaults to "darkblue".}

\item{colour_t2tINV}{Colour of the arcs representing tail-to-tail inversions (t2tINV). Defaults to "black".}

\item{colour_TRA}{Colour of the arcs representing translocations (TRA). Defaults to "darkgray".}

\item{color_minor_cn}{Colour for the horizontal bars representing the minor copy number values. Defaults to "#8491B4B2".}

\item{upper_limit_karyotype}{Upper limit in y-axis for karyotype ideogram. Defaults to -0.2 .}

\item{karyotype_rel_size}{Size of the karyotype ideogram, relative to the CN y axis. Defaults to .2.}

\item{custom_annotation}{Dataframe with custom annotation. Must contain columns "chr", "pos" and "y". Defaults to NULL.}

\item{ann_dot_col}{Colour of the dots in annotation plot. Defaults to "0.5."black".}

\item{ann_dot_size}{Size of the dots in annotation plot. Defaults to 0.5.}

\item{ann_y_title}{Label for Y-axis in annotation plot. Defaults to "".}

\item{ann_rel_size}{Size of the annotation plot, relative to the main plot. Defaults to 0.4.}

\item{curvature_intrachr_SVs}{Curvature for the arcs represeting intrachromosomal SVs. Defaults to -0.15.}

\item{curvature_interchr_SVs}{Curvature for the arcs represeting interchromosomal SVs. Defaults to -0.08.}

\item{max.cn}{Cap on the total copy number. Defaults to 8.}

\item{npc_now}{Controls spacing, needed to find SV partners on different chromosomes in the plot. Not recommended to change. Defaults to .00625 * 3.}

\item{scale_ticks}{Spacing of breaks in the x axis (in bp). Defaults to 20000000 (i.e., 20Mb).}

\item{size_interchr_SV_tip}{Size of the line indicating interchromosomal SVs involving chromosomes not shown (that is, not included in chr_selection).}

\item{label_interchr_SV}{Flag to annotate the second chromosome of interchromosomal SVs involving chromosomes not shown. Defaults to FALSE.}

\item{size_sv_line}{Linewidth for SVs. Defaults to .1.}

\item{genome_version}{Reference genome used. Can be either hg19, hg38, T2T, mm10 or mm39. Defaults to "hg38".}
}
\description{
This function generates publication-quality plots for copy number and structural variation profiles, which are particularly useful in the context of cancer genome analysis projects.
}
